# swagger.yml
openapi: 3.0.0
info:
  title: Swagger amazon clone - OpenAPI 3.0
  description: |-
    termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://amazonclone3.swagger.io/api/v3
...
paths:
...
  /product/all:
    get:
      tags:
        - product
      summary: Returns all products based filter
      operationId: getProducts
      parameters:
        - name: perPage
          in: query
          description: Per page data should be fetched
          required: true
          schema:
            type: number
        - name: page
          in: query
          required: true
          description: Page of data should be fetched
          schema:
            type: number
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid input
  /product/{id}:
    get:
      tags:
        - product
      summary: Returns  product details based on productID
      operationId: getProductDetails
      parameters:
        - name: id
          in: path
          description: productId to fetch the details
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid input
...